Move the query keys to src/utility/query
It makes the most sense for them to be located there.
