+++ apps/front-end/src/resources/Blogs/queries/useBlogQuery.ts
@@ -4,7 +4,7 @@ import { parseBlogView } from "@lexicon/models";
 
 import useQuery from "src/hooks/query/useQuery";
 import lexiconAuthenticatedClient from "src/utility/lexiconAuthenticatedClient";
-import queryKeys from "src/utility/queryKeys";
+import queryKeys from "src/utility/query/queryKeys";
 
 function useBlogQuery(blogId: string, params?: BlogQueryStringInputType) {
   return useQuery<BlogView>({
