+++ apps/front-end/src/resources/Blogs/queries/useBlogRevisionsQuery.ts
@@ -4,7 +4,7 @@ import { parseBlogRevisionHistory } from "@lexicon/models";
 
 import useQuery from "src/hooks/query/useQuery";
 import lexiconAuthenticatedClient from "src/utility/lexiconAuthenticatedClient";
-import queryKeys from "src/utility/queryKeys";
+import queryKeys from "src/utility/query/queryKeys";
 
 function useBlogRevisionsQuery(blogId: string) {
   return useQuery<Array<BlogRevision>>({
