+++ apps/front-end/src/resources/Users/queries/useUpdateUserProfileMutation.ts
@@ -4,7 +4,7 @@ import { useQueryClient } from "@tanstack/react-query";
 
 import useMutation from "src/hooks/query/useMutation";
 import lexiconAuthenticatedClient from "src/utility/lexiconAuthenticatedClient";
-import queryKeys from "src/utility/queryKeys";
+import queryKeys from "src/utility/query/queryKeys";
 
 function useUpdateUserProfileMutation() {
   const queryClient = useQueryClient();
